Stopped working here | Check description
I stopped working here because I encountered a problem when I tried to add a recycler view for both rest and exercises activities, the problem was that the rv need the Exercise List, which is an object, and when one activity is switched to another this object is reinitialized every time. I tried to move the object with Intent, but the app get too complicated. Probably this issue can be fixed using coroutines, so I decided to start over from the point I split the app into two activities and implement again everything into a single activity.